Create TracksViewController
