+++ RideTogether/RideTogether/MapViewDelegate.swift
@@ -86,34 +86,4 @@ class MapViewDelegate: NSObject, MKMapViewDelegate {
             })
         }
     }
-    // MARK: 選擇路線後導航
-    func mapView(_ mapView: MKMapView, didSelect view: MKAnnotationView) {
-        guard let annotation = view.annotation else {
-                return
-            }
-
-            let directionRequest = MKDirections.Request()
-            directionRequest.source = MKMapItem.forCurrentLocation()
-            directionRequest.destination = MKMapItem(placemark: MKPlacemark(coordinate: annotation.coordinate))
-            directionRequest.transportType = .automobile
-            let directions = MKDirections(request: directionRequest)
-
-            directions.calculate {
-                (response, error) -> Void in
-                guard let response = response else {
-                    if let error = error {
-                        print("Error: \(error)")
-                    }
-                    return
-                }
-
-                if !response.routes.isEmpty {
-                    let route = response.routes[0]
-                    DispatchQueue.main.async { [weak self] in
-//                        self?.mapView.add(route.polyline)
-                    }
-                }
-            }
-    }
-    
 }
