+++ RideTogether/RideTogether/View/Profile.storyboard
@@ -35,14 +35,14 @@
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="7i6-aU-asl">
-                                <rect key="frame" x="0.0" y="88" width="414" height="725"/>
+                                <rect key="frame" x="0.0" y="381" width="414" height="432"/>
                                 <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                             </tableView>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="vDu-zF-Fre"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
-                            <constraint firstItem="7i6-aU-asl" firstAttribute="top" secondItem="vDu-zF-Fre" secondAttribute="top" id="0j4-cl-VF3"/>
+                            <constraint firstItem="7i6-aU-asl" firstAttribute="top" secondItem="vDu-zF-Fre" secondAttribute="top" constant="293" id="0j4-cl-VF3"/>
                             <constraint firstItem="vDu-zF-Fre" firstAttribute="bottom" secondItem="7i6-aU-asl" secondAttribute="bottom" id="WQy-qR-gTy"/>
                             <constraint firstItem="7i6-aU-asl" firstAttribute="leading" secondItem="vDu-zF-Fre" secondAttribute="leading" id="aD2-LO-m2y"/>
                             <constraint firstItem="vDu-zF-Fre" firstAttribute="trailing" secondItem="7i6-aU-asl" secondAttribute="trailing" id="bPR-xa-cQk"/>
@@ -52,7 +52,22 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="Ief-a0-LHa" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="-430" y="50"/>
+            <point key="canvasLocation" x="-474" y="50"/>
+        </scene>
+        <!--View Controller-->
+        <scene sceneID="TC8-NO-X9b">
+            <objects>
+                <viewController id="Y0Y-w7-fwb" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="ttQ-6p-dhQ">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <viewLayoutGuide key="safeArea" id="d2B-FO-ic1"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="pZV-P3-N87" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="276.81159420289856" y="49.553571428571423"/>
         </scene>
     </scenes>
     <resources>
