+++ RideTogether/RideTogether/ViewController/Profile/ProfileViewController.swift
@@ -15,5 +15,4 @@ class ProfileViewController: UIViewController {
         // Do any additional setup after loading the view.
     }
     
-
 }
